requirements.txt
Flask==2.1.2
Flask-Bootstrap5==0.5.0
Flask-SQLAlchemy==2.5.1
